Doc hover fix to repl links with plain text
